change 5
I punch through all the code because i didn't want to continue wasting
hours trying to figure out what I did wrong. In the end i tried to run
the code but had errors with layout. I know where the problem is because
it was my the second time i got stuck after the layout issue of not
being recognized. I am going back to the book and re look over the
instructions